pyproject.toml
[project]
name = "kk-ocr-v2"
version = "2.0.0"
description = "Pipeline Ekstraksi 'Kartu Keluarga' V2 (YOLO + U-Net + VLM)"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {text = "MIT"}
authors = [
    {name = "KK-OCR Team"}
]
keywords = ["ocr", "computer-vision", "deep-learning", "fastapi", "yolo", "unet", "vlm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core API Framework
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    # Deep Learning & Computer Vision
    "torch==2.1.1",
    "torchvision==0.16.1",
    "ultralytics==8.0.221",
    "segmentation-models-pytorch==0.3.3", # Pretrained U-Net for enhancement
    "opencv-python==4.8.1.78",
    "pillow==10.1.0",
    "numpy==1.26.2",
    "scikit-image==0.22.0",
    # VLM Integration
    "google-generativeai==0.3.1",
    "requests==2.31.0",
    # Authentication & Security
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.0",
    # Monitoring & Logging
    "prometheus-client==0.19.0",
    "structlog==23.2.0",
    # Utilities
    "pyyaml==6.0.1",
    "click==8.1.7",
    "tqdm==4.66.1",
    "jsonschema==4.20.0",
    # Data Processing
    "pandas==2.1.3",
    "albumentations==1.3.1",
    "gradio>=4.19.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "httpx==0.25.2",
    
    # Development Tools
    "black==23.11.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "isort==5.12.0",
    "pre-commit==3.5.0",
]

model-management = [
    "dvc==3.30.1",
]

all = [
    "kk-ocr-v2[dev,model-management]",
]

[project.urls]
Homepage = "https://github.com/yourusername/kk-ocr-v2"
Documentation = "https://github.com/yourusername/kk-ocr-v2/blob/main/README.md"
Repository = "https://github.com/yourusername/kk-ocr-v2"

[project.scripts]
kk-ocr = "src.api.main:main"
train-unet = "src.training.train_unet:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
# UV dependency groups (replaces dev-dependencies)
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.2",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = ["venv", ".venv", "build", "dist"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=src --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
